[serve] Deflake test_cluster (#39002)
I wasn't able to reproduce the flakiness locally, but based on the tracebacks, seems like in `test_node_failure` when the node is killed, we'll still try to send requests to the dead actor before controller health checks it and realizes it's dead. This PR simply considers that and ignores `RayActorError`s raised from `ray.get` in the test.